#173 moving towards isolating treasury concepts in econ.py
